Separate the gulpfiles
