+++ .gitignore
@@ -31,6 +31,7 @@ node_modules
 
 !typings
 app/**/*.js
+!app/gulpfile.js
 test/app/*.js
 test/app/instrumented
 dist
