+++ Gulpfile.js
@@ -1,3 +1,4 @@
+require('./app/gulpfile'); // load app's tasks
 var path = require('path');
 var del = require('del');
 var gulp = require('gulp'); //Gulp 4.0+
@@ -14,59 +15,17 @@ var lint = require('gulp-tslint');
 var gutil = require('gulp-util');
 var tslint = require('tslint');
 
-var lintConfig = require('./tslint.json');
-var lintReporter = function(output, file, options) {
-  output.forEach(function(failure) {
-    gutil.log(failure.name+"["+failure.startPosition.line+", "+failure.startPosition.character+"] ("+failure.ruleName+"): "+failure.failure);
-  });
-};
-
 var coverageFile = './coverage/coverage.json';
-
-function makeFrontendBuildTarget(name, projectfile) {
-  var project = ts.createProject(projectfile, {typescript: typescript});
-  gulp.task(name, gulp.series(function() {
-    return project.src()
-      .pipe(sourcemaps.init())
-      .pipe(ts(project))
-      .pipe(sourcemaps.write())
-      .pipe(gulp.dest('./tmp')); // Write every JS file to tmp/
-  }, function() {
-    return gulp.src('./tmp/**/*.js')
-      .pipe(sourcemaps.init())
-      .pipe(webpack({
-        entry: './tmp/main.js',
-        output: {
-          filename: 'app.js'
-        },
-        node: {
-          fs: 'empty'
-        },
-        module: {
-          loaders: [
-            { test: /\.json/, loader: 'json' },
-          ]
-        }
-      }))
-      .pipe(sourcemaps.write())
-      .pipe(gulp.dest('./dist')); // webpack everything to dist/app.js
-    }, function(cb) {
-      del('tmp').then(function() { cb(); });
-    }, function() {
-      return gulp.src('./app/index.html')
-        .pipe(gulp.dest('./dist'));
-    }));
-    gulp.task('lint:'+name, function() {
-      var tsFiles = gulpFilter(['**/*.ts', '!**/*.d.ts']);
-      return project.src()
-        .pipe(tsFiles)
-        .pipe(lint({configuration: lintConfig, tslint: tslint}))
-        .pipe(lint.report(lintReporter, {emitError: false}));
-    });
-}
 
 var testDest = './test/app';
 function makeTestBuildTarget(name, projectfile) {
+  var lintConfig = require('./app/tslint.json');
+  var lintReporter = function(output, file, options) {
+    output.forEach(function(failure) {
+      gutil.log(failure.name+"["+failure.startPosition.line+", "+failure.startPosition.character+"] ("+failure.ruleName+"): "+failure.failure);
+    });
+  };
+
   var project = ts.createProject(projectfile, {typescript: typescript});
   gulp.task(name+'-build', function() {
     return project.src()
@@ -90,10 +49,9 @@ function makeTestBuildTarget(name, projectfile) {
   });
 }
 
-makeFrontendBuildTarget('build:app', 'app/tsconfig.json');
 makeTestBuildTarget('build:test:app', 'test/app/tsconfig.json');
 
-gulp.task('lint', gulp.parallel('lint:build:app', 'lint:build:test:app'));
+gulp.task('lint', gulp.parallel('lint:app', 'lint:build:test:app'));
 
 gulp.task('build:test', gulp.parallel('build:test:app'));
